Extract Ui and Parser class
