+++ src/main/java/Duke/Commands/DoneTaskCommand.java
@@ -2,8 +2,9 @@ package Duke.Commands;
 
 import Duke.Duke;
 import Duke.Task.InvalidTaskException;
-import Duke.Task.Task;
 import Duke.Task.TaskList;
+import Duke.Ui.Ui;
+import Duke.Ui.UserInput;
 
 import java.util.HashSet;
 import java.util.List;
@@ -14,11 +15,11 @@ class DoneTaskCommand extends Command {
     private static final String DONE_TASK_SUCCESS_MESSAGE = "Good job! I've marked this task as done:\n\t%s";
 
     @Override
-    public void run(Duke duke, Duke.UserInput input) throws InvalidTaskException {
+    public void run(Duke duke, UserInput input) throws InvalidTaskException {
         int taskIndex = parseTaskNumber(input);
         TaskList taskList = duke.getTaskList();
         taskList.markAsDone(taskIndex);
-        duke.say(String.format(DONE_TASK_SUCCESS_MESSAGE, taskList.get(taskIndex)));
+        Ui.print(String.format(DONE_TASK_SUCCESS_MESSAGE, taskList.get(taskIndex)));
     }
 
     @Override
