+++ src/main/java/Duke/Commands/ExitCommand.java
@@ -1,6 +1,7 @@
 package Duke.Commands;
 
 import Duke.Duke;
+import Duke.Ui.UserInput;
 
 import java.util.HashSet;
 import java.util.List;
@@ -10,7 +11,7 @@ class ExitCommand extends Command {
     private static final Set<String> KEYWORDS = new HashSet<>(List.of("exit", "bye", "quit"));
 
     @Override
-    public void run(Duke duke, Duke.UserInput input) {
+    public void run(Duke duke, UserInput input) {
         duke.stop();
     }
