+++ src/main/java/Duke/Commands/RemoveTaskCommand.java
@@ -4,6 +4,8 @@ import Duke.Duke;
 import Duke.Task.InvalidTaskException;
 import Duke.Task.Task;
 import Duke.Task.TaskList;
+import Duke.Ui.Ui;
+import Duke.Ui.UserInput;
 
 import java.util.HashSet;
 import java.util.List;
@@ -15,11 +17,11 @@ class RemoveTaskCommand extends Command {
             + TASKS_COUNT_MESSAGE;
 
     @Override
-    public void run(Duke duke, Duke.UserInput input) throws InvalidTaskException {
+    public void run(Duke duke, UserInput input) throws InvalidTaskException {
        int taskIndex = parseTaskNumber(input);
        TaskList taskList = duke.getTaskList();
        Task task = taskList.remove(taskIndex);
-       duke.say(String.format(REMOVE_TASK_SUCCESS_MESSAGE, task, taskList.size()));
+       Ui.print(String.format(REMOVE_TASK_SUCCESS_MESSAGE, task, taskList.size()));
     }
 
     @Override
